Commit 2 of 4 for bringing the changes over from the hetero branch. Merged in from:
svn merge -r5448:5496 https://svn.open-mpi.org/svn/ompi/tmp/hetero .

This commit was SVN r5550.

The following SVN revisions from the original message are invalid or
inconsistent and therefore were not cross-referenced:
  r5448
  r5496

+++ src/mca/gpr/proxy/gpr_proxy_general_operations.c
@@ -26,7 +26,7 @@
 #include "orte_config.h"
 
 #include "include/orte_constants.h"
-#include "dps/dps_types.h"
+#include "mca/dps/dps_types.h"
 #include "util/output.h"
 #include "util/proc_info.h"
 
@@ -36,7 +36,7 @@
 
 #include "gpr_proxy.h"
 
-int orte_gpr_proxy_preallocate_segment(char *name, int num_slots)
+int orte_gpr_proxy_preallocate_segment(char *name, size_t num_slots)
 {
     return ORTE_ERR_NOT_IMPLEMENTED;
 }
